requirements.txt
gspread==6.1.2
google-auth==2.29.0
google-generativeai==0.5.4
requests==2.31.0
python-dotenv==1.0.1
